fix: rename args into flags to avoid confusion
