+++ cmd/fullstack-network-manager/src/commands/cluster/index.mjs
@@ -1,11 +1,11 @@
 import {Cluster} from './cluster.mjs'
-import * as cmdArgs from '../../core/args.mjs'
+import * as flags from '../../core/flags.mjs'
 
 const createSubCmd = {
     command: 'create',
     desc: 'Create FST cluster',
     builder: yargs => {
-        yargs.option('name', cmdArgs.clusterName)
+        yargs.option('name', flags.clusterName)
     },
     handler: argv => {
         Cluster.create(argv)
@@ -16,7 +16,7 @@ const deleteSubCmd = {
     command: 'delete',
     desc: 'Delete FST cluster',
     builder: yargs => {
-        yargs.option('name', cmdArgs.clusterName)
+        yargs.option('name', flags.clusterName)
     },
     handler: argv => {
         Cluster.delete(argv)
