+++ cmd/fullstack-network-manager/src/core/flags.mjs
